SMXS-37: PR comments, discussions, changes

+++ src/httpDto/TaskResponse.go
@@ -2,5 +2,5 @@ package httpDto
 
 type TaskResponse struct {
 	Id       int64 `json:"id"`
-	QueuedAt int   `json:"queued_at"`
+	QueuedAt int64 `json:"queued_at"`
 }
